add undeploy flow #44

+++ devops/maven/dew-maven-plugin/src/main/java/com/tairanchina/csp/dew/kernel/flow/unrelease/DefaultUnReleaseFlow.java
@@ -20,14 +20,32 @@ import com.tairanchina.csp.dew.helper.KubeHelper;
 import com.tairanchina.csp.dew.kernel.Dew;
 import com.tairanchina.csp.dew.kernel.flow.BasicFlow;
 import io.kubernetes.client.ApiException;
+import io.kubernetes.client.models.V1Pod;
+import io.kubernetes.client.models.V1beta1ReplicaSet;
 import org.apache.maven.plugin.MojoExecutionException;
 
 import java.io.IOException;
+import java.util.List;
 
 public class DefaultUnReleaseFlow extends BasicFlow {
 
     public boolean process() throws ApiException, IOException, MojoExecutionException {
-        KubeHelper.delete(Dew.Config.getCurrentProject().getAppName(), Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.SERVICE, Dew.Config.getCurrentProject().getId());
+        if (KubeHelper.exist(Dew.Config.getCurrentProject().getAppName(), Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.SERVICE, Dew.Config.getCurrentProject().getId())) {
+            KubeHelper.delete(Dew.Config.getCurrentProject().getAppName(), Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.SERVICE, Dew.Config.getCurrentProject().getId());
+        }
+        if (KubeHelper.exist(Dew.Config.getCurrentProject().getAppName(), Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.DEPLOYMENT, Dew.Config.getCurrentProject().getId())) {
+            KubeHelper.delete(Dew.Config.getCurrentProject().getAppName(), Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.DEPLOYMENT, Dew.Config.getCurrentProject().getId());
+        }
+        List<V1beta1ReplicaSet> rsList = KubeHelper.list("app=" + Dew.Config.getCurrentProject().getAppName() + ",group=" + Dew.Config.getCurrentProject().getAppGroup() + ",version=" + Dew.Config.getCurrentProject().getAppVersion(),
+                Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.REPLICA_SET, V1beta1ReplicaSet.class, Dew.Config.getCurrentProject().getId());
+        for (V1beta1ReplicaSet rs : rsList) {
+            KubeHelper.delete(rs.getMetadata().getName(), rs.getMetadata().getNamespace(), KubeHelper.RES.REPLICA_SET, Dew.Config.getCurrentProject().getId());
+        }
+        List<V1Pod> pods = KubeHelper.list("app=" + Dew.Config.getCurrentProject().getAppName() + ",group=" + Dew.Config.getCurrentProject().getAppGroup() + ",version=" + Dew.Config.getCurrentProject().getAppVersion(),
+                Dew.Config.getCurrentProject().getNamespace(), KubeHelper.RES.POD, V1Pod.class, Dew.Config.getCurrentProject().getId());
+        for (V1Pod rs : pods) {
+            KubeHelper.delete(rs.getMetadata().getName(), rs.getMetadata().getNamespace(), KubeHelper.RES.POD, Dew.Config.getCurrentProject().getId());
+        }
         return true;
     }
